menambah database dan  user auth dan library kivyMd
menambha database bertipe noSql dengan library TyniDb dan yang di
kelolanya adalah file .json , selanjutnya menambah user untuk
autentikasi pada semua yang ada hubunganya dengan navigasi , library
__kivymd__ : untuk menambah icon pada button , icons , dan lainya degan
tujuan utnuk memperbagus GUI

+++ components/infoPopup.py
@@ -19,7 +19,7 @@ class InfoPopup:
 
     '''
 
-    def __init__(self, title: str, type: str, timer: int):
+    def __init__(self, title: str, type: str, timer: float):
         self.title = title
         self.type = type
         self.popup = None  # tentukan popup di awal
